requirements.txt
adal==1.2.7
appdirs==1.4.4
applicationinsights==0.11.10
appnope==0.1.2
APScheduler==3.9.1
argcomplete==2.0.0
argon2-cffi==21.1.0
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==19.3.0
autopep8==1.5.2
azure-common==1.1.28
azure-core==1.26.1
azure-cosmos==4.2.0
azure-functions==1.12.0
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==2.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-keyvault==10.1.0
azure-mgmt-resource==21.2.1
azure-mgmt-storage==20.1.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azureml-core==1.47.0
azureml-dataprep==4.5.7
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.11.4
azureml-dataset-runtime==1.47.0
azureml-opendatasets==1.47.0
azureml-telemetry==1.47.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0rc1
bcrypt==4.0.1
bleach==1.5.0
Brotli==1.0.9
bson==0.5.8
certifi==2022.12.7
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.2.0
contextlib2==21.6.0
cryptography==38.0.1
cycler==0.10.0
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.8.0
docker==6.0.1
dotnetcore2==3.1.23
entrypoints==0.3
Flask==2.1.2
Flask-Caching==1.10.1
Flask-Compress==1.12
Flask-Cors==3.0.8
flask-googlemaps==0.4.1
fusepy==3.0.1
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.3.0
ghp-import==2.1.0
greenlet==2.0.1
gunicorn==20.0.4
haversine==2.7.0
html5lib==0.9999999
htmlmin==0.1.12
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.18.0
jeepney==0.8.0
Jinja2==3.0.3
jmespath==1.0.1
jsmin==3.0.1
jsonpickle==2.2.0
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
knack==0.10.1
Levenshtein==0.20.8
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.1.0
marshmallow==3.5.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.4.0
mkdocs-material==5.5.14
mkdocs-material-extensions==1.0.3
mkdocs-minify-plugin==0.3.0
more-itertools==8.2.0
msal==1.20.0
msal-extensions==0.3.1
msrest==0.7.1
msrestazure==0.6.4
multitasking==0.0.11
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.22.3
oauthlib==3.2.2
overpy==0.6
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.2
pathlib-mate==1.1.1
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.9.2
plotly==5.7.0
pluggy==0.13.1
portalocker==2.6.0
prettytable==3.5.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.18.1
ptyprocess==0.7.0
py==1.8.1
py4j==0.10.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.6.0
pymdown-extensions==9.6
pymongo==3.7.1
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
pyspark==3.3.1
pytest==5.3.5
python-binary-memcached==0.29.0
python-dateutil==2.8.2
python-dotenv==0.12.0
python-Levenshtein==0.20.8
python-memcached==1.59
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
rapidfuzz==2.13.3
redis==4.4.0
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.9.3
SecretStorage==3.3.3
Send2Trash==1.8.0
six==1.14.0
SQLAlchemy==1.4.44
sqlalchemy-mate==1.4.28.3
tabulate==0.9.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
uhashring==1.1
urllib3==1.26.8
uszipcode==1.0.1
watchdog==2.1.9
wcwidth==0.1.8
websocket-client==1.4.2
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wrapt==1.14.1
yfinance==0.1.79
zipp==3.8.0
